@@ -1,5 +1,5 @@
 /*  Havok Format Library
-    Copyright(C) 2016-2020 Lukas Cone
+    Copyright(C) 2016-2022 Lukas Cone
 
     This program is free software : you can redistribute it and / or modify
     it under the terms of the GNU General Public License as published by
@@ -88,9 +88,4 @@ struct hkVirtualClass : IhkVirtualClass {
   virtual void Save(BinWritterRef wr, hkFixups &fixups) const;
 
   static hkVirtualClass *Create(JenHash hash, CRule rule);
-};
-
-template <class C> struct hkRealArray {
-  const C *data;
-  uint32 count;
 };
